Made JobBoard and Help Board. Created Exit
The job board and help board now work. There are some stuff that needs to be changed when we have more assets and ideas. Small things need to be changed as well with the help board returning to job board. Player can now leave and play game. Currently does not do any checks with maps selected or anything